+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+### 1.0.5
+
 ### 1.0.4
 + [x] Manipulation: optimise `append`, `prepend` to one definition for IE
 + [x] Script: add `$.script`, `$.fn.script` to execute scripts from element or create script object